+++ R/xvalidate.R
@@ -45,7 +45,7 @@
 #' unique_ids <- data.frame(id = id, fold = sample(1:nfolds, size = length(id), replace = TRUE))
 #' cross_ids <- merge(trade[, 5, drop = FALSE], unique_ids, by = "id", all.x = TRUE)
 #' # Finally, we try xvalidate with a lasso penalty (the default) and two lambda values:
-#' \donttest{reg <- xvalidate(y = y, x = x, fes = fes, lambda = 0.001,
+#' \dontrun{reg <- xvalidate(y = y, x = x, fes = fes, lambda = 0.001,
 #'                          IDs = cross_ids$fold, verbose = TRUE)}
 #'
 #' @inheritSection hdfeppml_int References
@@ -54,8 +54,7 @@ xvalidate <- function(y, x, fes, IDs, testID = NULL, tol = 1e-8, hdfetol = 1e-4,
                      init_mu = NULL, init_x = NULL, init_z = NULL, verbose = FALSE,
                      cluster = NULL, penalty = "lasso", method = "placeholder",
                      standardize = TRUE, penweights = rep(1, ncol(x_reg)), lambda = 0) {
-  message("start xval")
-  dim(IDs)
+
   ## incorporate folds option (removed from arguments for initial release, since it wasn't doing
   # anything. Create seed option.
   if (is.null(init_mu)) {
@@ -98,7 +97,7 @@ xvalidate <- function(y, x, fes, IDs, testID = NULL, tol = 1e-8, hdfetol = 1e-4,
   # print(length(IDs))
   #drop 1 id at a time and predict its mean values out of sample
   for (i in start_loop:n_IDs) {
-    print(i)
+#    print(i)
     omitID <- uniq_IDs[i]
     if(!is.null(testID)) {
       omitID <- testID
@@ -112,11 +111,6 @@ xvalidate <- function(y, x, fes, IDs, testID = NULL, tol = 1e-8, hdfetol = 1e-4,
       print(length(insample))
       print(length(y))
     }
-
-    print("length insample, length omitted, length y")
-    print(length(insample))
-    print(length(which(IDs==omitID)))
-    print(length(y))
 
     #select y and x
     y_temp   <- y[insample]